AVIServer/src/pb_to_json/requirements.txt
protobuf==3.1.0
paho-mqtt==1.2
pytz==2014.4
protobuf-to-dict=0.1.0
argparse==1.4.0
